coverage. Do not check raw logical expressions temporarily

+++ compiler/rustc_mir_build/src/errors.rs
@@ -823,8 +823,8 @@ pub struct NontrivialStructuralMatch<'tcx> {
 pub(crate) struct MCDCExceedsConditionNumLimit {
     #[primary_span]
     pub span: Span,
-    pub conditions_num: u16,
-    pub max_conditions_num: u16,
+    pub conditions_num: usize,
+    pub max_conditions_num: usize,
 }
 
 #[derive(Diagnostic)]
